requirements.txt
APScheduler==3.6.3
beautifulsoup4==4.11.1
bs4==0.0.1
cachetools==4.2.2
certifi==2022.9.14
html5lib==1.1
mechanize==0.4.8
numpy==1.23.3
pandas==1.5.0
python-dateutil==2.8.2
python-telegram-bot==13.14
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
schedule==1.1.0
six==1.16.0
soupsieve==2.3.2.post1
tornado==6.1
tzdata==2022.2
tzlocal==4.2
webencodings==0.5.1
